endpoints/requirements.txt
kafka-python
protobuf
flask==1.1.1
gunicorn==20.0.4